Added base Mapbase matchers to tier1 so that the response system library can access them
